pyproject.toml
[tool.poetry]
name = "investrss"
version = "0.1.0"
description = ""
authors = ["Ruslan Sirazhetdinov <ruslansir@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0.0"
tinkoff-investments = "<0.3.0"
rfeed = "^1.1.1"
fastapi = "^0.109.2"
punq = "^0.7.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
aiohttp = "^3.9.3"
ruff = "^0.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
